+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New functionality:
 * [OLMIS-6434](https://openlmis.atlassian.net/browse/OLMIS-6434): Calculated and populated Stock on Hand values to a new table.
 * [OLMIS-6483](https://openlmis.atlassian.net/browse/OLMIS-6483): Added saving calculated SoH value while sensing stock event and using it during getting stock card info.
 * [OLMIS-6558](https://openlmis.atlassian.net/browse/OLMIS-6558): Add new environment variable - PUBLIC_URL and use to for email generated links
+* [OLMIS-6573](https://openlmis.atlassian.net/browse/OLMIS-6573): Added possibility to fetch all valid sources/destinations via API.
 
 Improvements:
 * [OLMIS-5569](https://openlmis.atlassian.net/browse/OLMIS-5569): Moved reason TRANSFER_IN from demo data to bootstrap data.
@@ -46,7 +47,7 @@ Bug fixes, security and performance improvements, also backwards-compatible:
 ==================
 
 Contract breaking changes:
-* [OLMIS-4756](https://openlmis.atlassian.net/browse/OLMIS-4756): Can't change `type` or `category` for stock card line item reason on update. 
+* [OLMIS-4756](https://openlmis.atlassian.net/browse/OLMIS-4756): Can't change `type` or `category` for stock card line item reason on update.
 
 Improvements:
 * [OLMIS-4940](https://openlmis.atlassian.net/browse/OLMIS-4940): Ensured that the microservice gets system time zone from configuration settings on startup.